set the exec path more accurately
the exec path path in emacs did not always match the shell path, which
caused confusions because sometimes commands run with 'shell-command'.
this code (lifted from somewhere, probably stackoverflow) sets the exec
path from the output of running 'echo $PATH' in the shell. Note that
this only runs when the underlying OS is darwin (OSX), its reasonable to
think that it should also be ran on linux systems.