make zap-to-char end-char non-inclusive
e.g. if curser is '|':
'the qu|ck brown fox'
and `zap-to-char n` is called, the text that will be killed is 'ck
brown', but the text that is remaining will be (assuming the | was
covering an 'i'):
'the quin fox'.
Need to fix this so that the 'n' is not added to the kill ring...