add some snippets, a dired customization init file, secrets file
The secrets file is for holding api keys, passwords etc. The actual secrets file
should never be checked into source control, but whenever something is added to
the secrets file a placeholder version should be added to the
.private.example.el file for reference.
Add some dired customizations, and remove dired-details and dired-details+
packages as dependencies.